+++ src/serlo-editor/plugins/equations/editor-renderer.tsx
